+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Added `validateSessionUser()` to prevent that deleted or modified users could still access the backend until their session cookie expired.
+
 ## [0.2.0-alpha.12] - 2026-05-25
 
 ### Added